requirements.txt
gradio
openai
python-dotenv
pypdf
